pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux'" },
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
]

[project]
name = "adaptive-retrieval"
description = "Researching adaptive retrieval"
requires-python = ">=3.10"
dependencies = [
    "llama-index-llms-openai>=0.3.42",
    "chromadb==0.5.15",
    "autorag[gpu]>=0.3.13",
    "click>=8.1.8",
    "kiwipiepy>=0.20.4",
    "numpy==1.26.4",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.4",
    "setuptools>=78.1.0",
    "setuptools-scm>=8.2.0",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "pytorch-lightning>=2.5.1.post0",
    "wandb>=0.19.11",
    "pytest-dotenv>=0.5.2",
    "llama-index-llms-openai>=0.3.42"
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = { file = ["src/VERSION"] }

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests"]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = ["E722", "F821"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = ["B"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "tab"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
